pyproject.toml
[project]
name = "template-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "wemake-python-styleguide>=1.1.0",
]
